requirements.txt
base==0.0.0
Django==5.2.1
django-allauth==65.10.0
djangorestframework==3.16.0
gunicorn==23.0.0
matplotlib==3.10.5
mpmath==1.3.0
numpy==2.2.6
pandas==2.2.3
pillow==11.2.1
python-dateutil==2.9.0.post0
typing_extensions==4.13.2
